+++ Lib/Rendering/World.cpp
@@ -28,7 +28,7 @@ namespace rendering {
 
     void World::clearObjects() {
         // Delete all objects and clear the vector
-        for (int i = 0; i < objects.size(); ++i) {
+        for (size_t i = 0; i < objects.size(); ++i) {
             if (objects[i] != nullptr) {
                 delete objects[i];
             }
